feat(NonRegistrantsFormUI): Moved generic form folder to be a child of the src folder
