+++ src/GenericRegistrationRequestForm/GenericRegistrationRequestForm.css
@@ -1,0 +1,59 @@
+
+.generic-reg-container {
+    padding: 30px 32px;
+  }
+
+  .generic-reg-form-container {
+    border-radius: 8px;
+    background-color: #fff;
+    padding: 10px 16px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    font-family: inherit;
+  }
+
+  .generic-reg-form {
+    width: 100%;
+  }
+
+  .generic-reg-exp-text {
+    width: 100%;
+    display: flex;
+    justify-content: flex-end;
+  }
+
+  @media screen and (min-width: 1280px) {
+    .generic-reg-form {
+      width: 60%;
+      min-width: 50%;
+    }
+  }
+
+  .generic-reg-form-item {
+    display: flex;
+    align-items: baseline;
+  }
+
+  .generic-reg-form-item__middle-icon {
+    padding-left: 8px;
+  }
+
+  .generic-reg-form-item__last-item {
+    padding-left: 8px;
+    flex-grow: 1;
+    flex-shrink: 0;
+  }
+
+  .generic-reg-disclaimer-text {
+    font-size: 0.8rem;
+    text-align: center
+  }
+
+
+
+.generic-reg-disclaimer-text {
+    font-size: 0.8rem;
+    text-align: center
+  }
